+++ src/javascripts/Helpers/printToDom.js
@@ -1,0 +1,6 @@
+const printToDom = (divId, textToPrint) => {
+  const selectedDiv = document.querySelector(divId);
+  selectedDiv.innerHTML = textToPrint;
+};
+
+export default printToDom;
